Modify Util and Readme
